+++ src/muzero_gojax/train.py
@@ -115,7 +115,7 @@ def init_model(go_model: hk.MultiTransformed, absl_flags: absl.flags.FlagValues)
     """Initializes model either randomly or from laoding a previous save file."""
     rng_key = jax.random.PRNGKey(absl_flags.random_seed)
     if absl_flags.load_path:
-        params = load_params(absl_flags.load_path)
+        params = load_params(absl_flags.load_path, dtype='bfloat16')
         print(f"Loaded parameters from '{absl_flags.load_path}'.")
     else:
         params = go_model.init(rng_key, gojax.new_states(absl_flags.board_size, 1))
